Try to use more Sphinx code for parsing documents
I would like to use the same "code/rules" as Sphinx does, but it is not
clear if this is possible (since the Sphinx set up resolves around the
idea of an "app", which I don't want to set up here: maybe I should).

This has apparently tweaked some of the rules. I am also testing against
the latest version which has new docstrings in which may also have
contributed to some needed changes.