Extract ahelp metadata when run
This extends the previous commit to support extracting the ahelp
metadata at run-time to the doc2ahelp.py script, which means that
the unused first directory could be dropped.

Support passing the symbol itself to the routine generating the
ahelp file. This is to support advanced functionality - e.g. adding
an example for all models showing the default parameters - but it
is not used.