Try to use existing ahelp metadata for multi-command files
The old system had files such as 'ahelp fit' which described both
fit and fit_bkg. This was done by including fit_bkg in the
refkeywords of the fit file. We can therefore identify (hopefully)
these files by looking at the mapping from refkeywords to key
(i.e. in this case fit_bkg -> fit).

We can then use the metadata from the file (fit in this example)
to fill out the fit_bkg file (which is new).