Create an AHELP file for synonyms
Previously I was not creating an ahelp file for synonyms, such
as covariance, as you could just just use covar. Unfortunately these
names are used enough in the refkeywords that 'ahelp covariance'
just returned a lot of matches rather than 'ahelp covar', so now
we create ahelp files for them.

I am not 100% convinced the seealsogroup stuff is sensible here,
as you now get covar and covariance listed, but I am not sure it's
worth fixing.