Better handle CIAO 4.17 files
This is mainly tracking the many new XSPEC models added in this
release.

The existing code tries to catch the notes sections to remove
the "needs XSPEC blah" text we have for the standalone version.
It looks like this can be cleaned up but I have not tried to do
too much here.