show_ahelp now works with annotations and annotation changes
The show_ahelp.py CLI will now use almost the same logic as
doc2ahelp.py but just for a single file, so it makes it easier
to check when there's a problem.

The layout of the annotated signature has been broken out by
argument, and more effort has been made to use "modern" (> python 3.9)
forms, even when the code uses them.

Not 100% sure about all this.