Do not include annotated signature in the PARAMETERS section
For CIAO 4.17 it feels like the annotated signature - i.e. the
signature with types - is not helpful, as the information is
already included in the parameter table.

There are a few cases where it does add extra information, but
not enough to warrant the "this information is duplicated" issue.

This can be reviewed for CIAO 4.18.